pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "market-observability"
version = "0.2.0"
description = "Market observability system with anomaly detection and AI-enhanced reporting"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "feedparser>=6.0",
    "pydantic>=2.0",
    "typer[all]>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "yfinance>=0.2.30",
    "jinja2>=3.1",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
